+++ cx-portal-shared-components/src/components/basic/StaticTemplate/Cards/ImageTextSideBySide.tsx
@@ -19,6 +19,7 @@
  ********************************************************************************/
 
 import { Box } from '@mui/material'
+import { ImageItem } from '../../ImageGallery/ImageItem'
 import { Typography } from '../../Typography'
 import { ProviderProps } from '../StaticTypes'
 
@@ -37,7 +38,17 @@ export default function ImageTextSideBySide({
         padding: '90px 0px',
       }}
     >
-      <img src={provider.imageUrl} width="482" alt={'alt tag info'} />
+      <ImageItem
+        url={provider.imageUrl || ''}
+        text={''}
+        size="custom"
+        height="472px"
+        width="100%"
+        hover={true}
+        borderRadius={true}
+        shadow={false}
+        modalWidth="1100"
+      />
       <Box
         sx={{
           padding: '20px',