+++ CHANGELOG.md
@@ -8,6 +8,8 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
    * BugFix - Register daps via the auth icon in the connector registration, allows all type of files
 * Notification
    * BugFix - Wrong url link
+* Static Templates
+   * Show image on modal for a wider view
 * ImageGallery
    * Rectangle images to be 16:9 ratio.
    * Fix 3 images to a row. Show grid layout for rest